More modulue reorganizataion around xxxValues
UpValues moved to properites. Allows a string parameter now as well.
DRY eval_xxxValue code.
