+++ SYMBOLS_MANIFEST.txt
@@ -636,7 +636,6 @@ System`KelvinKei
 System`KelvinKer
 System`Key
 System`KeyDropFrom
-System`Key
 System`KeyExistsQ
 System`Keys
 System`Khinchin
